==> 1_2_get_data.Rmd <==
---
title: "get_data"
author: "ipi"
date: "3/27/2020"
output: html_document
---

```{r setup2, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

check.packages <- function(pkg){
  new.pkg <- pkg[!(pkg %in% installed.packages()[, "Package"])]
  if (length(new.pkg))
    install.packages(new.pkg, dependencies = TRUE,
                     repos = "https://cran.rstudio.com")
  sapply(pkg, require, character.only = TRUE)
}

pkgList <- c("RCurl","WDI","readxl","httr","knitr","tidyverse",
             "estimatr","stargazer","countrycode","dataverse",
             "zip", "lubridate", "rvest", "fastDummies", "lfe",
             "zoo","rvest","data.table", "pbapply", "xml2")

check.packages(pkgList)
```

```{r, eval = !exists("download")}
download <- list()
download$corona_data      <- FALSE
download$wdi_data         <- FALSE
download$qog_data         <- FALSE
download$jhu_data         <- FALSE
download$polity_data      <- FALSE
download$swiid_data       <- FALSE
download$resp_data        <- FALSE
download$popden_data      <- FALSE
download$epr_data         <- FALSE
download$who_data         <- FALSE
download$weather_data     <- FALSE
download$dpi_data         <- FALSE
download$parlgov_data     <- FALSE
download$wiki_data        <- FALSE
download$ipu_data         <- FALSE
download$iefs_data        <- FALSE
download$mobility_data    <- FALSE
download$stringency_data  <- FALSE

# If `corona_data` is predefined in `0_master`, use this value
if(exists("corona_data")) download$corona_data <- corona_data
```

* Get Corona Data from ECDC (This data gives cases and deaths per day, not cumulative. Below we cumulate and calculate growth rates.) 

```{r download corona data,  message = FALSE, warning = FALSE}

if(download$corona_data){
  
  url <- paste("https://www.ecdc.europa.eu/sites/default/files/documents/COVID-19-geographic-disbtribution-worldwide-",format(Sys.time(), "%Y-%m-%d"), ".xlsx", sep = "")
  
  # url <- paste("https://www.ecdc.europa.eu/sites/default/files/documents/COVID-19-geographic-disbtribution-worldwide-",format(Sys.time(), "2020-06-16"), ".xlsx", sep = "")
  
  GET(url, authenticate(":", ":", type="ntlm"), write_disk(tf <- tempfile(fileext = ".xlsx")))
  
  write.csv(read_excel(tf), "saved/corona_download.csv", row.names = F)
}

```

## Pull data

```{r load corona data, message = FALSE, warning = FALSE}
corona <- read.csv("saved/corona_download.csv") %>%
      rename(geoid           = geoId,
             cases           = cases,
             deaths          = deaths,
             date_rep        = dateRep,
             population_2018 = popData2018,
             continent       = continentExp,
             country         = countriesAndTerritories,
             country_id      = countryterritoryCode)%>% 
       mutate(date = as.Date(date_rep)) %>% 
  dplyr::select(-c(year,month,day,geoid))

# Correct manually 2 cases without a ISO3 code
corona <- corona %>% 
  mutate(country = as.character(country),
         country_id = as.character(country_id)) %>% 
  mutate(country_id = if_else(country=="Anguilla", "AIA", country_id),
         country_id = if_else(country=="Bonaire, Saint Eustatius and Saba", "BES", country_id),
         country_id = if_else(country=="Falkland_Islands_(Malvinas)", "FLK", country_id))

# 2-letter "geoid" should be avoided due to Namibia (NA) problem
rectangle <- expand_grid(unique(corona$country_id), seq(as.Date(min(corona$date)),
                                                    as.Date(max(corona$date)),
                                                    by = "day"))
names(rectangle) <- c("country_id", "date")

rectangle <- rectangle %>% 
  mutate(month = format(as.POSIXct(date), "%m"),
         day = format(as.POSIXct(date), "%d"),
         year = format(as.POSIXct(date), "%Y"),
         elapsed = as.integer(-1 + date - min(date)))

corona <- left_join(rectangle, corona) %>% 
  arrange(country_id, elapsed) %>% 
  mutate(deaths = as.numeric(deaths),
         cases = as.numeric(cases),
         deaths = if_else(is.na(deaths), 0, deaths),
         cases  = if_else(is.na(cases), 0, cases)) %>% 
  group_by(country_id) %>%
  
  mutate(cases_cum      = cumsum(cases),
         deaths_cum     = cumsum(deaths),
         deaths_cum_log = log(1+deaths_cum),
         deaths_cum_l7  = lag(deaths_cum, n = 7, order_by = elapsed),
         deaths_cum_g7  = lag(deaths_cum/deaths_cum_l7 - 1, order_by = elapsed),
         deaths_cum_g7  = ifelse(deaths_cum_g7 == Inf, NA, deaths_cum_g7)) %>%
  ungroup() %>% 
  
  mutate(country      = as.character(country),
         country_id   = as.character(country_id),
         country_id   = if_else(country_id == "XKX", "RKS", country_id)) %>% 
  filter(!is.na(country_id)) # Remove "Western Sahara"

# Plug in missing values for country names, since some merging is done based on
# this variable.
.mode <- function(x) {
  ux <- unique(x)
  ux[which.max(tabulate(match(x, ux)))]
}

corona <- corona %>% 
  group_by(country_id) %>% 
  mutate(country = .mode(na.omit(country))) %>% 
  ungroup()

write.csv(corona, "saved/corona_df.csv")
```

## Get JHU data

```{r load/clean JHU, message = FALSE, eval = download$jhu_data}

# HT reshape code from https://www.r-bloggers.com/tidying-the-new-johns-hopkins-covid-19-time-series-datasets/

confirmed_raw <- read_csv("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv")
deaths_raw <- read_csv("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_global.csv")



.jhd_to_long <- function(df) {
  df_str <- deparse(substitute(df))
  var_str <- substr(df_str, 1, str_length(df_str) - 4)
  
  df %>% group_by(`Country/Region`) %>%
    filter(`Country/Region` != "Cruise Ship") %>%
    select(-`Province/State`, -Lat, -Long) %>%
    mutate_at(vars(-group_cols()), sum) %>% 
    distinct() %>%
    ungroup() %>%
    rename(country = `Country/Region`) %>%
    pivot_longer(
      -country, 
      names_to = "date_str", 
      values_to = var_str
    ) %>%
    mutate(date = mdy(date_str)) %>%
    select(country, date, !! sym(var_str)) 
}

jh_covid19_data <- .jhd_to_long(confirmed_raw) %>%
  full_join(.jhd_to_long(deaths_raw))

```

## Get WDI data

```{r download wdi, eval = download$wdi_data}

# write.csv(WDI_data$series,  "saved/wdi_series.csv")

wdi <- WDI(
  country = "all", 
  indicator = c(
    gov_effect  = "GE.EST", 
    trade       = "NE.TRD.GNFS.ZS",
    ineq        = "SI.DST.10TH.10", 
    gdp_pc      = "NY.GDP.PCAP.PP.KD",
    pop_tot     = "SP.POP.TOTL",
    older_m     =  "SP.POP.65UP.MA.IN",
    older_f     = "SP.POP.65UP.FE.IN",
    air_travel  = "IS.AIR.PSGR",
    fdi         = "BX.KLT.DINV.CD.WD",
    pop_density = "EN.POP.DNST",
    urban       = "EN.URB.MCTY.TL.ZS"), 
  start = 2018, end = 2018, extra = FALSE, cache = NULL)

# Migration share from 2015

wdi <- left_join(wdi, WDI(country = "all", 
         indicator = c(migration_share = "SM.POP.TOTL.ZS"), 
  start = 2015, end = 2015, extra = FALSE, cache = NULL) %>% 
  select(iso2c, migration_share))

# Oil from 2017

wdi <- left_join(wdi, WDI(country = "all", 
         indicator = c(oil = "NY.GDP.PETR.RT.ZS"), 
  start = 2017, end = 2017, extra = FALSE, cache = NULL) %>% 
  select(iso2c, oil))

socin <- WDI(country = "all", 
             indicator = c(soc_insur_cov ='PER_SI_ALLSI.COV_POP_TOT'),
             start = 2010, end = 2018) %>% 
  filter(!is.na(soc_insur_cov)) %>% 
  group_by(country) %>% 
  arrange(year) %>% 
  mutate(soc_insur_cov = last(soc_insur_cov)) %>% ungroup() %>% 
  distinct(country, .keep_all = T)

wdi <- left_join(wdi, socin %>% dplyr::select(iso2c, soc_insur_cov))

## Social contributions

soccon <- WDI(country = "all", 
              indicator = c(soc_contrib ='GC.REV.SOCL.ZS'),
              start = 2010, end = 2018) %>% 
  filter(!is.na(soc_contrib)) %>% 
  group_by(country) %>% 
  arrange(year) %>% 
  mutate(soc_contrib = last(soc_contrib)) %>% ungroup() %>% 
  distinct(country, .keep_all = T)

wdi <- left_join(wdi, soccon %>% dplyr::select(iso2c, soc_contrib))

## Social safety

soc_safety <- WDI(country = "all", 
                  indicator = c(soc_safety ='PER_SA_ALLSA.COV_POP_TOT'),
                  start = 2010, end = 2018) %>% 
  filter(!is.na(soc_safety)) %>% 
  group_by(country) %>% 
  arrange(year) %>% mutate(soc_safety = last(soc_safety)) %>% ungroup() %>% 
  distinct(country, .keep_all = T)

wdi <- left_join(wdi, soc_safety %>% dplyr::select(iso2c, soc_safety)) %>%

  mutate(country_id = countrycode(iso2c,
                                    origin = "iso2c",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Netherlands Antilles", "ANT", country_id),
         country_id = if_else(country=="Namibia", "NAM", country_id),
         country_id = if_else(country=="Kosovo", "RKS", country_id)) %>% 
    dplyr::select(-c(year, iso2c, country)) %>% 
  filter(!(is.na(country_id)))

write.csv(wdi, "saved/wdi_df.csv")

```


## Get Quality of Government data


```{r get-qog-data, eval = download$qog_data}
qogDF <- read.csv(url("http://www.qogdata.pol.gu.se/data/qog_std_cs_jan20.csv"),
                  header = TRUE) %>% 
  mutate(country = as.character(cname)) %>% 
  dplyr::select(country, vdem_libdem, al_ethnic2000,
                al_religion2000, fe_etfra, vdem_mecorrpt) %>% 
  rename(al_etfra = al_ethnic2000,
         al_religfra = al_religion2000)

# Recoding.
qogDF <- qogDF %>% 
  mutate(country = recode(country, 
    "Brunei"= "Brunei Darussalam",
    "Cape Verde" = "Cabo Verde",
    "Taiwan" = "Taiwan, China",
    "Congo" =  "Congo, Rep.",
    "Congo, Democratic Republic" =  "Congo, Dem. Rep.",
    "Cyprus (1975-)" =  "Cyprus",
    "Ethiopia (1993-)" =  "Ethiopia",
    "France (1963-)" =  "France",
    "Gambia" =  "Gambia, The",
    "Iran" =  "Iran, Islamic Rep.",
    "Korea, South" =  "Korea, Rep.",
    "Kyrgyzstan" =  "Kyrgyz Republic",
    "Malaysia (1966-)" =  "Malaysia",
    "Pakistan (1971-)" =  "Pakistan",
    "Russia" =  "Russian Federation",
    "St Vincent and the Grenadines" =  "St. Vincent and the Grenadines",
    "Slovakia" =  "Slovak Republic",
    "Sudan (2012-)" =  "Sudan",
    "Syria" =  "Syrian Arab Republic",
    "Egypt" =  "Egypt, Arab Rep.",
    "Venezuela" =  "Venezuela, RB")) %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Micronesia", "FSM", country_id)) %>% 
  dplyr::select(-country)

write.csv(qogDF,  file = "./saved/qog_df.csv", row.names = FALSE)
```



## Get Polity IV data

```{r get-polity-data, eval = download$polity_data}
download.file("http://www.systemicpeace.org/inscr/p4v2018.xls", 
              "./saved/p4v2018.xls",
              mode = "wb",
              quiet = TRUE)
polityDF <- read_excel(path = "./saved/p4v2018.xls",
                       col_names = TRUE)

# Cleanup and ID recoding
polityDF <- polityDF %>% 
  dplyr::select(scode, year, polity2) %>% 
  filter(year==2018) %>% 
  mutate(
    country_id = countrycode(sourcevar = scode, origin = "p4c", destination = "iso3c"),
    country_id = if_else(scode=="KOS", "RKS", country_id)) %>% 
  rename(polity = polity2) %>% 
  dplyr::select(-c(scode,year))

write.csv(polityDF,
          file = "./saved/polity4_df.csv",
          row.names = FALSE)
file.remove("./saved/p4v2018.xls")
```

## Get SWIID inequality data


```{r get-swiid-data, eval = download$swiid_data}
download.file("https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/LM4OWF/DBWK5H", 
              "./saved/swiid8_2.zip",
              mode = "wb",
              quiet = TRUE)

unzip("./saved/swiid8_2.zip",
     files = "swiid8_2/swiid8_2_summary.csv",
     exdir = "./saved")
file.remove("./saved/swiid8_2.zip")

file.copy(from = "./saved/swiid8_2/swiid8_2_summary.csv",
          to = "./raw/swiid8_2_summary.csv")
file.remove("./saved/swiid8_2/swiid8_2_summary.csv")

swiidDF <- read.csv(file = "./raw/swiid8_2_summary.csv",
                    header = TRUE)

swiidDF <- swiidDF %>% 
  dplyr::select(country, year, gini_disp) %>% 
  rename(gini = gini_disp) %>% 
  filter(year>=2010) %>% 
  group_by(country) %>% 
  summarise(gini = mean(gini, na.rm = TRUE)) %>% 
  mutate(country_id = countrycode(sourcevar = country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Kosovo", "RKS", country_id)) %>% 
  mutate(country_id = if_else(country=="Micronesia", "FSM", country_id)) %>% 
  mutate(country_id = if_else(country=="SÃ£o TomÃ© and PrÃ­ncipe", "STP", country_id)) %>% 
  dplyr::select(-c(country))

write.csv(swiidDF, file = "./saved/swiid_df.csv", row.names = FALSE)
file.remove("./raw/swiid8_2_summary.csv")
```


```{r declare policy response categories}
lockdown <- c('Partial lockdown', 'Full lockdown')
distancing <- c('Limit public gatherings', 'Schools closure', 'Public services closure', 'Curfews')
```



```{r download policy responses, warning=F,message=F, eval = download$resp_data}
# download.file("https://www.acaps.org/sites/acaps/files/resources/files/20200330_acaps_-_covid-19_goverment_measures_dataset_v3.xlsx# ", 
#               "./saved/20200330-acaps-covid-19-goverment-measures-dataset-v3.xlsx",
#               mode = "wb",
#               quiet = TRUE)
# 
# download.file("https://www.acaps.org/sites/acaps/files/resources/files/20200423_acaps_-_covid-19_goverment_measures_dataset_v10.xls# x", 
#               "./saved/acaps-v10.xlsx",
#               mode = "wb",
#               quiet = TRUE)

## Scrape ACAPS

page <- xml2::read_html("https://www.acaps.org/covid19-government-measures-dataset")
url_acaps <- page %>% 
  html_nodes(xpath = '//*[@id="inner-content"]/article/div[3]/div/div[2]/div/div/div/div/div[2]/div/span/a') %>% 
  html_attr('href')

## Download ACAPS

fname_acaps <- url_acaps %>% str_split(pattern = '/') %>% unlist() %>% .[length(.)]

GET(url_acaps, write_disk(tf <- tempfile(fileext = ".xlsx")))

write_csv(read_excel(tf, sheet = 4), "saved/acaps_current.csv")
```

```{r load/clean policy responses, eval = FALSE, message = FALSE}
resp <- read_csv('saved/acaps_current.csv') %>% 
  filter(is.na(ADMIN_LEVEL_NAME)) %>% ## only national policies 
  mutate(date = as.Date(DATE_IMPLEMENTED))

## Colnames to lower

colnames(resp) <- tolower(colnames(resp))

## Drop some more vars

resp <- resp %>% 
  dplyr::select(id, country, region, category, measure, date) %>% 
  mutate(measure = str_squish(measure))

## Need to create larger categories

lockdown <- c('Partial lockdown', 'Full lockdown')
distancing <- c('Limit public gatherings', 'Schools closure', 'Public services closure',
                'Curfews')

## Dummies for this

resp <- resp %>% 
  mutate(lockdown_measures = ifelse(measure %in% lockdown, 1, 0),
         distancing_measures = ifelse(measure %in% distancing, 1, 0)) %>% 
  group_by(country, region, date) %>% 
  summarise(lockdown_measures = sum(lockdown_measures, na.rm = T),
            distancing_measures = sum(distancing_measures, na.rm = T)) %>% 
  ungroup() %>% 
  mutate(distancing_bin = ifelse(distancing_measures > 0 ,1, 0),
         lockdown_bin = ifelse(lockdown_measures > 0, 1, 0))

## Merge with corona structure
cor_merge <- read.csv("saved/corona_df.csv") %>% 
  dplyr::select(geoid, country, date_rep) %>% 
  filter(!is.na(date_rep) & !is.na(country)) %>% 
  mutate(date = as.Date(as.character(date_rep))) %>% 
  dplyr::select(-date_rep) 

## Merge

cor_merge <- cor_merge %>% 
  left_join(resp)%>% 
  arrange(country, date) %>% 
  mutate_at(vars(lockdown_measures, distancing_measures),
            list(~ifelse(is.na(.), 0, .)))

## Get cum sum of measures, gen dummies

cor_merge <- cor_merge %>% 
  group_by(country) %>% 
  mutate(lockdown_n = cumsum(lockdown_measures),
         distancing_n = cumsum(distancing_measures),
         distancing_bin = ifelse(distancing_n > 0 ,1, 0),
         lockdown_bin = ifelse(lockdown_n > 0 ,1, 0)) %>% 
  dplyr::select(-matches('measures'))

## Days since first distancing / lock down

cor_merge_policy <- cor_merge %>% 
  group_by(country) %>% 
  mutate(days_rel_ldown = date - date[lockdown_bin == 1][1],
         days_rel_dstng = date - date[distancing_bin == 1][1]) %>% 
  ungroup() %>% 
  dplyr::select(geoid, date, matches('distancing|lockdown'))

# Azerbaijan has 2 observations for "February 29", producing a duplicate in
# the Covid-19 data.  
cor_merge_policy <- cor_merge_policy %>% 
  slice(-which(cor_merge_policy$geoid=="AZ" & cor_merge_policy$date=="2020-02-29")[1])
  
rm(cor_merge, resp, distancing, lockdown)
# Switch to 3-letter code for countries
cor_merge_policy <- cor_merge_policy %>% 
  mutate(country_id = countrycode(geoid,
                                    origin = "iso2c",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(geoid=="XK", "RKS", country_id),
         country_id = if_else(geoid=="EL", "GRC", country_id),
         country_id = if_else(geoid=="UK", "GBR", country_id)) %>% 
  dplyr::select(-geoid)

write.csv(cor_merge_policy, "saved/acaps_df.csv")
# No correspondence could be found for code "JPG11668"

## This could now be merged to the main DF
```

## Pop density data (World Bank)


```{r pop density (second time), warning=F, message=F}
pop_den <- read_csv('raw/pop_den_worldbank.csv', skip = 4) %>% 
  dplyr::select(`Country Name`, `2018`) %>% 
  rename(country = 1, pop_den_2018 = 2) %>% 
  filter(!is.na(pop_den_2018)) %>% 
  mutate(country = recode(country, `Lao PDR` = 'Laos'))

# Create 3-letter code for country
pop_den <- pop_den %>% 
  mutate(country_id = countrycode(country, origin = "country.name", destination = "iso3c"),
         country_id = if_else(country=="Kosovo", "RKS", country_id)) %>% 
  dplyr::select(-country) %>%
  data.frame()


write.csv(pop_den, "saved/pop_density_df.csv")

```

## Life expectancy (source is WB)

```{r life expectancy, warning=F, message=F, eval = FALSE}
life_expect <- read_csv('raw/life_expectancy.csv', skip = 4) %>% 
  dplyr::select(`Country Name`, `2017`) %>% 
  rename(country = 1, life_exp_2017 = 2) %>% 
  filter(!is.na(life_exp_2017)) %>% 
  mutate(country = recode(country, `Lao PDR` = 'Laos'))

# Create 3-letter code for country
life_expect <- life_expect %>% 
  mutate(country_id = countrycode(country, origin = "country.name", destination = "iso3c"),
         country_id = if_else(country=="Kosovo", "RKS", country_id)) %>% 
  dplyr::select(-country)

write.csv(life_expect, 'saved/life_exp_df.csv')
```

## Pop below 14 (source is WB)

```{r pop below 14, warning=F, message=F, eval = FALSE}

# NOT PROPERLY SOURCED

pop_below_14 <- read_csv('raw/pop_age_below14.csv', skip = 4) %>% 
  dplyr::select(`Country Name`, `2018`) %>% 
  rename(country = 1, pop_below14_2018 = 2) %>% 
  filter(!is.na(pop_below14_2018)) %>% 
  mutate(country = recode(country, `Lao PDR` = 'Laos'))

# Create 3-letter code for country
pop_below_14 <- pop_below_14 %>% 
  mutate(country_id = countrycode(country, origin = "country.name", destination = "iso3c"),
         country_id = if_else(country=="Kosovo", "RKS", country_id)) %>% 
  dplyr::select(-country)

write.csv(pop_below_14, "saved/pop_below14_df.csv")
```

## Median age (source is WHO)

```{r median age who, warning=F, message=F}
med_age <- read_csv('raw/median_age_who.csv') %>% 
  rename(country = 1, med_age_2013 = 2, bla = 3) %>% 
  dplyr::select(-3) %>% 
  slice(-1) %>% 
  mutate(country = recode(country,
                          `Bolivia (Plurinational State of)` = 'Bolivia',
                          `Congo` = 'Congo, Rep.',
                          `Czechia` = 'Czech Republic',
                          `Democratic Republic of the Congo` = 'Congo, Dem. Rep.',
                          `United States of America` = 'United States',
                          `United Kingdom of Great Britain and Northern Ireland` = 'United Kingdom',
                          `Egypt` = 'Egypt, Arab Rep.',
                          `Iran (Islamic Republic of)` = 'Iran, Islamic Rep.',
                          `Venezuela (Bolivarian Republic of)` = 'Venezuela, RB',
                          `Republic of Korea` = 'Korea, Rep.',
                          `Democratic People's Republic of Korea` = 'Korea, North',
                          `Slovakia` = 'Slovak Republic',
                          `Kyrgyzstan` = 'Kyrgyz Republic',
                          `Viet Nam` = 'Vietnam',
                          `Gambia` = 'Gambia, The',
                          `United Republic of Tanzania` = 'Tanzania'))

# Create 3-letter code for country
med_age <- med_age %>% 
  mutate(country_id = countrycode(country, origin = "country.name", destination = "iso3c"),
         country_id = if_else(country=="Eswatini", "SWZ", country_id)) %>% 
  dplyr::select(-country)

write.csv(med_age, "saved/median_age_df.csv")

```

## Global health security index

```{r clean HGI, warning=F, message=F}
ghi <- read_csv('raw/global health security index raw.csv') %>% 
  t() %>% data.frame(stringsAsFactors = F)
colnames(ghi) <- ghi[1, ]
ghi$country <- row.names(ghi)

# Remove top 3 rows
ghi <- ghi %>% 
  slice(4:n())

## Select some vars
keep <- c('6.5.2b', '6.5.3a', '6.2.1a', '4.1.2a', 
          '4.1.1a', '6.5.1b', 'BG4', 'country')
keep_labels <- c('access_sanitation', 'health_exp_capita', 
                 'adult_lit_rate', 'hosp_beds_capita', 
                 'doctors_capita','healthcare_qual_index',
                 'hdi')

removeFUN <- function(x) {
  str_remove(x, ",")
}
replaceFUN <- function(x) {
  str_replace(x, "-", NA_character_)
}

# Have to recode health expenditure per capita, since the
# thousands separator in the numbers are leading to NAs being
# created.
ghi <- ghi %>% 
  dplyr::select(matches(paste0(keep, collapse = '|'))) %>% 
  dplyr::select(-matches('field epi|airlines|capacity|country vulner')) %>% 
  rename(doctors_pc = 1,
         hosp_beds_pc = 2,
         literacy_rate = 3,
         healthcare_qual = 4,
         acc_sanitation = 5,
         health_exp_pc = 6,
         hdi = 7) %>% 
  mutate_at(vars(-country), removeFUN) %>% 
  mutate_at(vars(-country), replaceFUN) %>%
  mutate_at(vars(-country), as.numeric) %>% 
  filter(!is.na(country))

## Recode some countries
ghi <- ghi %>% 
  mutate(country = recode(country, 
                          `Congo (Democratic Republic)` = 'Congo, Dem. Rep.',
                          `South Korea` = 'Korea, Rep.',
                          `Syria` = 'Syrian Arab Republic',
                          `Venezuela` = 'Venezuela, RB',
                          `Slovakia` = 'Slovak Republic',
                          `Egypt` = 'Egypt, Arab Rep.',
                          `Iran` = 'Iran, Islamic Rep.',
                          `Russia` = 'Russian Federation',
                          `Gambia` = 'Gambia, The',
                          `Congo (Brazzaville)` = 'Congo, Rep.'))

ghi$country[grep("Ivoire", ghi$country)] <- "Cote d'Ivoire"
ghi$country[grep("ncipe", ghi$country)] <- "Sao Tome and Principe"

ghi <- ghi %>% 
  mutate(country_id = countrycode(country, 
                                  origin = "country.name", 
                                  destination = "iso3c"), 
         country_id = if_else(country=="Micronesia", "FSM", country_id)) %>% 
  dplyr::select(-country)

write.csv(ghi, "saved/ghsi_df.csv")
```

* Pandemic preparedness

```{r get pandemic preparedness, warning = F, message=F}
## NEEDS SOURCING
panprep <- read_csv('raw/pandemic_preparedness.csv', 
                    col_names = F) %>% 
  rename(country = 2, pandemic_prep = 3) %>% 
  dplyr::select(-1) %>% 
  mutate(country = recode(country, 
                          "Congo Democratic Republic" = "Congo, Dem. Rep.",
                          "South Korea" = "Korea, Rep.",
                          "Syria" = "Syrian Arab Republic",
                          "Venezuela" = "Venezuela, RB",
                          "Slovakia" = "Slovak Republic",
                          "Egypt" = "Egypt, Arab Rep.",
                          "Iran" = "Iran, Islamic Rep.",
                          "Russia" = "Russian Federation",
                          "Gambia" = "Gambia, The",
                          "Congo Brazzaville" = "Congo, Rep."))

panprep$country[grep("Ivoire",panprep$country)] <- "Cote d'Ivoire"
panprep$country[grep("ncipe", panprep$country)] <- "Sao Tome and Principe"
panprep <- panprep %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Micronesia", "FSM", country_id)) %>% 
  dplyr::select(-country)

write.csv(panprep, "saved/pandemic_prep_df.csv")
```

* Health protection coverage

```{r get health protection coverage, warning = F, message=F}
health_prot_cov <- read.csv('raw/health-protection-coverage.csv', 
                            header = TRUE) %>% 
  rename(country = 1, share_health_insurance = 4) %>% 
  dplyr::select(-2, -3) %>% 
  filter(!is.na(country)) %>% 
  mutate(country = recode(country, 
                          `Democratic Republic of Congo` = 'Congo, Dem. Rep.',
                          `South Korea` = 'Korea, Rep.',
                          `Syria` = 'Syrian Arab Republic',
                          `Venezuela` = 'Venezuela, RB',
                          `Slovakia` = 'Slovak Republic',
                          `Egypt` = 'Egypt, Arab Rep.',
                          `Iran` = 'Iran, Islamic Rep.',
                          `Russia` = 'Russian Federation',
                          `Gambia` = 'Gambia, The',
                          `Congo Brazzaville` = 'Congo, Rep.',
                          `Kyrgyzstan` = 'Kyrgyz Republic',
                          `Macedonia` = 'North Macedonia')) %>% 
  mutate(country_id = countrycode(country, 
                                  origin = "country.name", 
                                  destination = "iso3c")) %>% 
  dplyr::select(-country) %>% 
  rename(share_health_ins = share_health_insurance)

write.csv(health_prot_cov, "saved/share_health_ins.csv")
```

* State Fragility Index

```{r get state fragility index data, warning=F, message=F}
fragility_index <- read_csv('raw/fragility_index_2018.csv') %>% 
  mutate(country = recode(country,
                          "Korea, South" = "Korea, Rep.",
                          "Dem. Rep. of Congo" = "Congo, Dem. Rep.",
                          "Congo-Brazzaville" = "Congo, Rep.",
                          "Egypt" = "Egypt, Arab Rep.",
                          "Iran" = "Iran, Islamic Rep.",
                          "Russia" = "Russian Federation",
                          "Venezuela" = "Venezuela, RB",
                          "Syria" = "Syrian Arab Republic",
                          "Kyrgyzstan" = "Kyrgyz Republic",
                          "Laos" = "Lao PDR",
                          "Macedonia" = "North Macedonia",
                          "Timor Leste" = "Timor-Leste",
                          "Sudan (North)" = "Sudan",
                          "Gambia" = "Gambia, The"))

# Switch to 3-letter country code
fragility_index <- fragility_index %>% 
  mutate(country_id = countrycode(country, origin = "country.name", destination = "iso3c"),
         country_id = if_else(country=="Kosovo", "RKS", country_id)) %>% 
  rename(state_fragility = sfi) %>% 
  dplyr::select(-c(country, iso2code))

write.csv(fragility_index, "saved/state_fragility_df.csv")
```  
  
 
* EPR ethnic fractionalization

```{r get-III-ethnic-fractionalization, warning = F, message=F, eval = download$epr_data}
eprCoreDF <- read.csv(url("https://icr.ethz.ch/data/epr/core/EPR-2019.csv"), header = TRUE)

# Basic cleanup
eprCoreDF <- eprCoreDF %>% 
  dplyr::select(gwid, statename, to, group, size, status) %>% 
  rename(country = statename,
         year = to) %>% 
  group_by(country) %>% 
  filter(year==max(year)) %>% 
  dplyr::select(-year) %>% 
  ungroup()

# From this, we can easily compute a Herfindahl concentration formula
# Create a residual group that brings up population to 100%
elfDF <- eprCoreDF %>% 
  mutate(country = as.character(country),
         group = as.character(group),
         status = as.character(status)) %>%
  group_by(gwid) %>%
  do(add_row(., .after = 0)) %>% 
  ungroup() %>%
  mutate(gwid = na.locf(gwid, fromLast = TRUE),
         country = na.locf(country, fromLast = TRUE),
         group = if_else(is.na(group), "Rest of country", group),
         status = if_else(is.na(status), "ARTIFICIALLY ADDED", status)) %>% 
  group_by(country) %>%
  mutate(size = if_else(is.na(size), 1-sum(size, na.rm = TRUE), size)) %>% 
  dplyr::select(-gwid) %>% 
  summarise(elf_epr = 1 - sum(size^2),
            rq_polarization = 4*sum(size^2*(1-size)),
            count_powerless = sum(status=="POWERLESS"),
            share_powerless = sum(size[status=="POWERLESS"]))

# Do a further small recoding to get an ad-hoc index of inclusion.
elfDF <- elfDF %>% 
  filter(!(country %in% c("Germany Democratic Republic",
                          "Czechoslovakia",
                          "Republic of Vietnam",
                          "Serbia and Montenegro",
                          "Zanzibar",
                          "Yemen People's Republic"))) %>%
  mutate(country_id = countrycode(country, 
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Kosovo", "RKS", country_id),
         country_id = if_else(country=="Tibet", "TIB", country_id)) %>% 
  dplyr::select(-country)


write.csv(elfDF,
          file = "./saved/eps_df.csv",
          row.names = FALSE)
```

* Global disease burden 2017: respiratory disease prevalence

```{r global burden of diseases: respiratory}
gbd17 <- read_csv('raw/gdb_respiratory.csv') %>% 
  filter(measure_name == 'Prevalence') %>% 
  filter(str_detect(cause_name, 'respiratory')) %>% 
  filter(metric_name == 'Percent') %>% 
  group_by(location_name) %>% 
  summarise(val = sum(val, na.rm = T)) %>% 
  mutate(val = val * 100) %>% 
  rename(country = location_name, resp_disease_prev = val)  %>% 
  mutate(country = recode(country, 
                          "Democratic Republic of the Congo" = "Congo, Dem. Rep.",
                          "South Korea" = "Korea, Rep.",
                          "Syria" = "Syrian Arab Republic",
                          "Venezuela" = "Venezuela, RB",
                          "Slovakia" = "Slovak Republic",
                          "Egypt" = "Egypt, Arab Rep.",
                          "Iran" = "Iran, Islamic Rep.",
                          "Russia" = "Russian Federation",
                          "The Gambia" = "Gambia, The",
                          "Congo" = "Congo, Rep.",
                          "Kyrgyzstan" = "Kyrgyz Republic",
                          "Macedonia" = "North Macedonia"))

# Switch to 3-letter country code
gbd17 <- gbd17 %>% 
  mutate(country_id = countrycode(country, 
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  dplyr::select(-country)

write.csv(gbd17, "saved/respiratory_disease_df.csv")

```

* Experience with epidemics (WHO)

```{r experience-with-epidemics, eval = download$who_data}
#### SARS ####
webpage <- read_html("https://www.who.int/csr/sars/country/table2004_04_21/en/")
# See how many tables are in the page
tbls <- html_nodes(webpage, "table")
head(tbls)
rm(tbls)
# We only need the first table there
tbls_ls <- webpage %>%
  html_nodes("table") %>%
  .[1] %>%
  html_table(fill = TRUE)
# Just from a brief examination, that table is quite ugly, so it will need
# some cleaning up before we can even think about merging it.
sarsDF <- as.data.frame(tbls_ls[[1]]); rm(tbls_ls)
sarsDF <- sarsDF %>%
  slice(-(1:2)) %>%
  dplyr::select(X1, X4, X6) %>%
  rename(country = X1,
         infections = X4,
         deaths = X6) %>%
  mutate(infections = str_remove(infections, fixed("^b")),
         infections = str_remove(infections, fixed("^c")),
         infections = as.numeric(infections),
         deaths = as.numeric(deaths)) %>%
  slice(1:(n()-2))
rm(webpage)
write.csv(sarsDF,
          file = "./raw/sars_df.csv",
          row.names = FALSE)
#### MERS ####
# Unfortunately, the "tabulizer" does not seem to help with
# much in this situation.
# Have to read the "raw" data directly from the hard drive
# Information on deaths is only found on Wikipedia, but this
# was considered unreliable
mersDF <- read.csv(file = "./raw/WHO-MERS.csv",
                   header = TRUE)
#### Ebola ####
ebolaDF <- read.csv(url("https://data.humdata.org/dataset/0d089fa0-3567-4b01-9c03-39d340ff34e3/resource/c59b5722-ca4b-41ca-a446-472d6d824d01/download/ebola_data_db_format.csv"),
                    header = TRUE)
ebolaDF <- ebolaDF %>%
  mutate(Indicator = as.character(Indicator),
         Country = as.character(Country),
         Date = as.character(Date),
         value = as.numeric(value)) %>%
  filter(Indicator %in% c("Cumulative number of confirmed, probable and suspected Ebola cases",
                          "Cumulative number of confirmed Ebola cases",
                          "Cumulative number of confirmed, probable and suspected Ebola deaths",
                          "Cumulative number of confirmed Ebola deaths")) %>%
  filter(Date=="2016-03-23")
# Clean up country names
ebolaDF <- ebolaDF %>% 
  dplyr::select(-Date) %>%
  mutate(Country = recode(Country,
                          "Liberia 2" = "Liberia",
                          "Guinea 2" = "Guinea")) %>%
  group_by(Country, Indicator) %>%
  summarise(value = sum(value)) %>%
  ungroup()
# Recoding
ebolaDF <- ebolaDF %>%
  mutate(Indicator = recode(Indicator,
                            "Cumulative number of confirmed Ebola cases" = "confirmedInfected",
                            "Cumulative number of confirmed Ebola deaths" = "confirmedDeaths",
                            "Cumulative number of confirmed, probable and suspected Ebola cases" = "totalInfected",
                            "Cumulative number of confirmed, probable and suspected Ebola deaths" = "totalDeaths")) %>% 
  pivot_wider(id_cols = Country, names_from = Indicator, values_from = value) %>%
  dplyr::select(Country, totalInfected, totalDeaths) %>% 
  rename(country = Country,
         infections = totalInfected,
         deaths = totalDeaths)
write.csv(ebolaDF,
          file = "./raw/ebola_df.csv",
          row.names = FALSE)
#### Merge the 3 data sets ####
mersDF <- mersDF %>%
  mutate(country = as.character(country),
         deaths = NA) %>% 
  rename(infections_mers = infections,
         deathsMERS = deaths)
sarsDF <- sarsDF %>%
  mutate(country = as.character(country)) %>% 
  rename(infections_sars = infections,
         deathsSARS = deaths) %>% 
  mutate(country = recode(country,
                          "United States" = "United States of America"))
ebolaDF <- ebolaDF %>%
  mutate(country = as.character(country)) %>% 
  rename(infections_ebola = infections,
         deathsEbola = deaths)
masterDF <- data.frame(country = unique(c(mersDF$country, 
                                          sarsDF$country, 
                                          ebolaDF$country)))
# Merging
masterDF <- left_join(masterDF, mersDF, by = "country") %>%
  left_join(sarsDF, by = "country") %>%
  left_join(ebolaDF, by = "country")
# Harmonize country names
masterDF <- masterDF %>%
  mutate(country = recode(country,
                          "Republic of Korea" = "South_Korea",
                          "Saudi Arabia" = "Saudi_Arabia",
                          "United Kingdom" = "United_Kingdom",
                          "United Arab Emirates" = "United_Arab_Emirates",
                          "United States of America" = "United_States_of_America",
                          "New Zealand" = "New_Zealand",
                          "Republic of Ireland" = "Ireland",
                          "Russian Federation" = "Russia",
                          "South Africa" = "South_Africa",
                          "Viet Nam" = "Vietnam",
                          "Sierra Leone" = "Sierra_Leone"))  %>%
  dplyr::select(country, infections_mers, infections_sars, infections_ebola) %>%
  mutate(infection = 0,
         infection = ifelse(!is.na(infections_mers) & infections_mers>100, 1, infection),
         infection = ifelse(!is.na(infections_sars) & infections_sars>100, 1, infection),
         infection = ifelse(!is.na(infections_ebola) & infections_ebola>100, 1, infection)) %>%
  # Switch to 3-letter country code
  mutate(country = recode(country,
                          "South_Korea" = "South Korea",
                          "United_States_of_America" = "United States")) %>% 
  mutate(country_id = countrycode(country, 
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  dplyr::select(-country)

write.csv(masterDF,
          file = "./saved/past_epidemics_df.csv",
          row.names = FALSE)
```

* Weather data

```{r weather-1, eval=download$weather_data}
#### TEMPERATURE DATA ####
# Measured in degrees Celsius
# Extract links from website, so as to be able to use them for data reading
page <- read_html("https://crudata.uea.ac.uk/cru/data/hrg/cru_ts_4.03/crucy.1905151143.v4.03/new_countries/tmp/")
linksVEC <- page %>% 
  html_nodes("a") %>% 
  html_attr("href")

linksVEC <- linksVEC[7:length(linksVEC)]
rm(page)

# Give full path for each link
linksVEC <- paste0("https://crudata.uea.ac.uk/cru/data/hrg/cru_ts_4.03/crucy.1905151143.v4.03/new_countries/tmp/",
                   linksVEC)
# Extract names of locations
namesVEC <- regmatches(linksVEC, 
                       regexpr("2018.\\s*\\K.*?(?=\\s*.tmp)", linksVEC, perl=TRUE))

# Custom function to compute a rolling mean without the current
# observation
.roll_without <- function(x, n) {
  # x is the vector to be transformed, while n is the count of
  # rows that need averaged
  require(zoo)
  xmean <- rollmeanr(x, n+1)
  xmean <- c(rep(NA, times = (length(x) - length(xmean))),
             xmean)
  xmeanright <- (xmean*(n+1) - x)/n
  return(xmeanright)
}

.extract_weather <- function(x) {
  # Produce list in which data sets are to be saved
  tempLIST <- list()
  for(i in 1:length(x)) {
    download.file(x[i],
                  "./raw/weatherTemp.per",
                  mode = "wb",
                  quiet = TRUE)
    per <- fread(file = "./raw/weatherTemp.per")
    per <- per %>%
      dplyr::select(YEAR:DEC) %>%
      pivot_longer(cols = JAN:DEC,
                   names_to = "month",
                   values_to = "temp") %>%
      filter(YEAR %in% c(2017,2018)) %>%
      mutate(temp = as.numeric(temp),
             temp = if_else(temp==-999, NA_real_, temp),
             temp_mean = .roll_without(temp, 3)) %>% 
      dplyr::select(-temp) %>% 
      filter(YEAR == 2018 | (YEAR==2017 & month=="DEC")) %>% 
      mutate(YEAR = YEAR + 2,
             country = namesVEC[i])
    
    tempLIST[[i]] <- per
  }
  bind_rows(tempLIST)
}

temperatureDF <- .extract_weather(linksVEC)
# Cleanup
file.remove("./raw/weatherTemp.per")


#### PRECIPITATION DATA ####
# Measured in mm/month
# Extract links from website, so as to be able to use them for data reading
page <- read_html("https://crudata.uea.ac.uk/cru/data/hrg/cru_ts_4.03/crucy.1905151143.v4.03/new_countries/pre/")
linksVEC <- page %>% 
  html_nodes("a") %>% 
  html_attr("href")

linksVEC <- linksVEC[7:length(linksVEC)]
rm(page)

# Give full path for each link
linksVEC <- paste0("https://crudata.uea.ac.uk/cru/data/hrg/cru_ts_4.03/crucy.1905151143.v4.03/new_countries/pre/",
                   linksVEC)
# Extract names of locations
namesVEC <- regmatches(linksVEC, 
                       regexpr("2018.\\s*\\K.*?(?=\\s*.pre)", linksVEC, perl=TRUE))

precipitationDF <- .extract_weather(linksVEC)
# Cleanup
file.remove("./raw/weatherTemp.per")
# Change name of variable
precipitationDF <- precipitationDF %>% 
  rename(precip = temp_mean)


#### MERGE AND CLEAN UP ####
weatherDF <- left_join(temperatureDF,
                       precipitationDF,
                       by = c("country","YEAR","month"))
rm(temperatureDF, precipitationDF)
write.csv(weatherDF,
          file = "./raw/weather_data_East_Anglia.csv",
          row.names = FALSE)
```

```{r weather-2}
weatherDF <- read.csv(file = "./raw/weather_data_East_Anglia.csv",
                      header = TRUE)

# Biggest challenge are country names
weatherDF <- weatherDF %>% 
  mutate(country = recode(country,
                          "Bosnia-Herzegovinia" = "Bosnia_and_Herzegovina",
                          "Brunei" = "Brunei_Darussalam",
                          "Cape_Verde_Isl" = "Cape_Verde",
                          "Central_African_Rep" = "Central_African_Republic",
                          "Curacao_Isl" = "Curacao",
                          "DR_Congo" = "Democratic_Republic_of_the_Congo",
                          "East_Timor" = "Timor_Leste",
                          "Faeroes" = "Faroe_Islands",
                          "Falkland_Isl" = "Falkland_Islands_(Malvinas)",
                          "Guinea-Bissau" = "Guinea_Bissau",
                          "Macedonia" = "North_Macedonia",
                          "Puerto_Rica" = "Puerto_Rico",
                          "Sao_Tome_+_Principe" = "Sao_Tome_and_Principe",
                          "St_Kitts_and_Nevis" = "Saint_Kitts_and_Nevis",
                          "St_Lucia" = "Saint_Lucia",
                          "St_Vincent" = "Saint_Vincent_and_the_Grenadines",
                          "Swaziland" = "Eswatini",
                          "Tanzania" = "United_Republic_of_Tanzania",
                          "USA" = "United_States_of_America",
                          "Virgin_Isl" = "United_States_Virgin_Islands"))%>% 
  rename(year = YEAR) %>% 
  mutate(month = recode(month,
                        "DEC" = "12",
                        "JAN" = "01",
                        "FEB" = "02",
                        "MAR" = "03",
                        "APR" = "04",
                        "MAY" = "05",
                        "JUN" = "06",
                        "JUL" = "07",
                        "AUG" = "08",
                        "SEP" = "09",
                        "OCT" = "10",
                        "NOV" = "11"),
         year = as.character(year))

# Switch to a 3-letter country code
weatherDF <- weatherDF %>% 
  mutate(country = str_replace_all(country, "_", " ")) %>% 
  filter(!(country=="Bassas da India")) %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Eswatini", "SWZ", country_id),
         country_id = if_else(country=="Guadalupe", "GLP", country_id),
         country_id = if_else(country=="Kosovo", "RKS", country_id),
         country_id = if_else(country=="Vanatu", "VUT", country_id)) %>% 
  dplyr::select(-country)

write.csv(weatherDF, "saved/weather_df.csv")
```

* V-DEM Indices - Media, equality of health system, accountability and rule of law

```{r read-vdem-data, eval=FALSE}
# Read in V-Dem data 
VDEMdf <- readRDS("./raw/V-Dem-CY-Full+Others-v10.rds")

# Subset of V-Dem data relevant for our analysis
# "v2meaccess" is no longer offered in the data as an indicator
VDEMdf <- VDEMdf %>% 
  group_by(country_name) %>% 
  filter(year==max(year, na.rm = TRUE)) %>%  
  ungroup() %>% 
  dplyr::select(country_name, v2mecrit, v2meharjrn, 
                v2pehealth, v2xcl_prpty, v2cltrnslw,
                v2x_pubcorr) %>%
  mutate(country_id = countrycode(country_name,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country_name=="Eswatini", "SWZ", country_id),
         country_id = if_else(country_name=="Kosovo", "RKS", country_id)) %>% 
  dplyr::select(-country_name) %>% 
  filter(!is.na(country_id)) %>% 
  rename(
    media_critical  = v2mecrit,
    journal_harass  = v2meharjrn,
    health_equality = v2pehealth,
    property_rights = v2xcl_prpty,
    transparent_law = v2cltrnslw,
    bureaucracy_corrupt = v2x_pubcorr)

# Palestinian territories appearing multiple times: (1) British Mandate;
# (2) Gaza Strip; (3) West Bank
# Somalia appearing only once, and was kept in the data set
VDEMdf <- filter(VDEMdf, !(country_id %in% c("PSE")))

# Write file in saved folder to merge 
write.csv(VDEMdf,
          file = "./saved/vdem_indexes_df.csv",
          row.names = FALSE)
```

```{r get prevent pandemics index data, warning=F, message=F}
prevent_df <- read_csv('raw/prevent_GHSI.csv') %>% 
  select(country, prevent_index) %>% 
  mutate(country = recode (country,
                            `South Korea` = 'Korea, Rep.',
                            `Congo (Democratic Republic)` = 'Congo, Dem. Rep.',
                            `Congo (Brazzaville)` = 'Congo, Rep.',
                            `Egypt` = 'Egypt, Arab Rep.',
                            `Iran` = 'Iran, Islamic Rep.',
                            `Russia` = 'Russian Federation',
                            `Venezuela` = 'Venezuela, RB',
                            `Syria` = 'Syrian Arab Republic',
                            `Laos` = 'Lao PDR',
                            `Gambia` = 'Gambia, The',
                            `Slovakia` = 'Slovak Republic',
                            `eSwatini (Swaziland)` = 'Eswatini'
                           )) 
prevent_df$country[grep("Ivoire", prevent_df$country)] <- "Cote d'Ivoire"
prevent_df$country[grep("ncipe", prevent_df$country)] <- "Sao Tome and Principe"

# Switch to 3-letter country code
prevent_df <- prevent_df %>% 
  mutate(country_id = countrycode(country,
                                  origin = "country.name",
                                  destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Eswatini", "SWZ", country_id),
         country_id = if_else(country=="Micronesia", "FSM", country_id)) %>%
  dplyr::select(-country)
```  

```{r get detect & report pandemics index data, warning=F, message=F}
detect_df <- read_csv('raw/detect_report_GHSI.csv') %>% 
  mutate(country = recode (country,
                            `South Korea` = 'Korea, Rep.',
                            `Congo (Democratic Republic)` = 'Congo, Dem. Rep.',
                            `Congo (Brazzaville)` = 'Congo, Rep.',
                            `Egypt` = 'Egypt, Arab Rep.',
                            `Iran` = 'Iran, Islamic Rep.',
                            `Russia` = 'Russian Federation',
                            `Venezuela` = 'Venezuela, RB',
                            `Syria` = 'Syrian Arab Republic',
                            `Laos` = 'Lao PDR',
                            `Gambia` = 'Gambia, The',
                           `Slovakia` = 'Slovak Republic',
                           `eSwatini (Swaziland)` = 'Eswatini'
                           )) 
detect_df$country[grep("Ivoire", detect_df$country)] <- "Cote d'Ivoire"
detect_df$country[grep("ncipe", detect_df$country)] <- "Sao Tome and Principe"

# Switch to 3-letter code for country
detect_df <- detect_df %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Micronesia", "FSM", country_id)) %>%
  dplyr::select(-country) %>% 
  rename(detect_index = detect_report_index)

write.csv(detect_df, "saved/ghsi_detect_df.csv")
```  

```{r get respond pandemics index data, warning=F, message=F}
respond_df <- read_csv('raw/respond_GHSI.csv') %>% 
  mutate(country = recode (country,
                            `South Korea` = 'Korea, Rep.',
                            `Congo (Democratic Republic)` = 'Congo, Dem. Rep.',
                            `Congo (Brazzaville)` = 'Congo, Rep.',
                            `Egypt` = 'Egypt, Arab Rep.',
                            `Iran` = 'Iran, Islamic Rep.',
                            `Russia` = 'Russian Federation',
                            `Venezuela` = 'Venezuela, RB',
                            `Syria` = 'Syrian Arab Republic',
                            `Laos` = 'Lao PDR',
                            `Gambia` = 'Gambia, The' ,
                           `Slovakia` = 'Slovak Republic',
                           `eSwatini (Swaziland)` = 'Eswatini'
                           ))
respond_df$country[grep("Ivoire", respond_df$country)] <- "Cote d'Ivoire"
respond_df$country[grep("ncipe", respond_df$country)] <- "Sao Tome and Principe"

# Switch to 3-letter code for country
respond_df <- respond_df %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Micronesia", "FSM", country_id)) %>%
  dplyr::select(-country)

write.csv(respond_df, "saved/ghsi_respond_df.csv")

```  

```{r get health index data, warning=F, message=F}
health_df <- read_csv('raw/health_GHSI.csv') %>% 
  mutate(country = recode (country,
                            `South Korea` = 'Korea, Rep.',
                            `Congo (Democratic Republic)` = 'Congo, Dem. Rep.',
                            `Congo (Brazzaville)` = 'Congo, Rep.',
                            `Egypt` = 'Egypt, Arab Rep.',
                            `Iran` = 'Iran, Islamic Rep.',
                            `Russia` = 'Russian Federation',
                            `Venezuela` = 'Venezuela, RB',
                            `Syria` = 'Syrian Arab Republic',
                            `Laos` = 'Lao PDR',
                            `Gambia` = 'Gambia, The',
                           `Slovakia` = 'Slovak Republic',
                           `eSwatini (Swaziland)` = 'Eswatini'
                           )) 
health_df$country[grep("Ivoire", health_df$country)] <- "Cote d'Ivoire"
health_df$country[grep("ncipe", health_df$country)] <- "Sao Tome and Principe"
               
# Switch to 3-letter code for country
health_df <- health_df %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Micronesia", "FSM", country_id)) %>%
  dplyr::select(-country)

write.csv(health_df, "saved/ghsi_health_df.csv")

```  

* Political polarization data from MARPOR

```{r get-marpor-data, eval=FALSE}
cmp <- read_excel(path = "./raw/MPDataset_MPDS2019b.xlsx", col_names = TRUE,
                  sheet = 1)

# Select latest year from each country
cmp <- cmp %>% 
  mutate(elect.yr = as.numeric(str_sub(edate, start = 1, end = 4))) %>% 
  group_by(countryname) %>% 
  filter(elect.yr==max(elect.yr, na.rm = TRUE)) %>%
  filter(elect.yr >= 2015) %>% 
  ungroup() %>% 
  dplyr::select(countryname, elect.yr, partyname, party,
                per503,per504,per505,per403,per404,per412,
                per401,per414,per601,per602,per603,per604,
                per605,per104,per201,per203,per305,per402,
                per407,per606,per103,per105,per106,per107,
                per406,per413,per506,per701,per702,per202,
                per604,per704,pervote)


# Load the South American version of the data set, and try to extract the same
# variables from it.
cmpsa <- read_excel(path = "./raw/MPDataset_MPDSSA2019b.xlsx", col_names = TRUE,
                    sheet = 1)

cmpsa <- cmpsa %>% 
  mutate(elect.yr = as.numeric(str_sub(edate, start = 1, end = 4))) %>% 
  group_by(countryname) %>% 
  filter(elect.yr==max(elect.yr, na.rm = TRUE)) %>%
  filter(elect.yr >= 2015) %>% 
  ungroup() %>% 
  dplyr::select(countryname, elect.yr, partyname, party,
                pervote,per503,per504,per505,per403,per404,
                per412,per401,per414,per601_1,per601_2,
                per602_1,per602_2,per603,per604,per605_1,
                per104,per201_1,per201_2,per203,per305_1,
                per305_2,per305_3,per305_4,per305_5,per305_6,
                per402,per407,per606_1,per606_2,per103_1,
                per103_2,per105,per106,per107,per406,
                per413,per506,per701,per702,per202_1,
                per202_3,per202_4,per604,per704) %>% 
  mutate(per601 = per601_1 + per601_2,
         per602 = per602_1 + per602_2,
         per201 = per201_1 + per201_2,
         per305 = per305_1 + per305_2 + per305_3 +
           per305_4 + per305_5 + per305_6,
         per606 = per606_1 + per606_2,
         per103 = per103_1 + per103_2,
         per202 = per202_1 + per202_3 + per202_4) %>%
  select(-c(per601_1, per601_2, per602_1, per602_2,
            per201_1, per201_2, per305_1, per305_2,
            per305_3, per305_4, per305_5, per305_6,
            per606_1, per606_2, per103_1, per103_2,
            per202_1, per202_3, per202_4)) %>% 
  rename(per605 = per605_1)

# Arrange columns to match those of the main CMP data set
cmpsa <- cmpsa[ ,names(cmp)]

cmp <- rbind(cmp, cmpsa); rm(cmpsa)

cmp <- cmp %>% 
  mutate_at(vars(per503:pervote), as.numeric)

# Examine which rows have a lot of NAs
as.data.frame(cmp[rowSums(is.na(cmp)) >= 5, 1:10])

# Remove the 8 parties listed, as some of them have not had
# their manifestos coded by the MARPOR project yet. A curious case is,
# for example "The River" (34340), which has a NA observation, as well
# as a second one, with full data availability
cmp <- cmp %>% 
  filter(!(countryname=="Denmark" & party==13410 & elect.yr==2015)) %>% 
  filter(!(countryname=="Iceland" & party==15630 & elect.yr==2017)) %>% 
  filter(!(countryname=="Greece" & party==34213 & elect.yr==2015)) %>% 
  filter(!(countryname=="Greece" & party==34340 & elect.yr==2015 & is.na(per503))) %>% 
  filter(!(countryname=="Greece" & party==34720 & elect.yr==2015 & is.na(per503))) %>% 
  filter(!(countryname=="Greece" & party==34730 & elect.yr==2015 & is.na(per503))) %>% 
  filter(!(countryname=="Croatia" & party==81022 & elect.yr==2016)) %>% 
  filter(!(countryname=="Slovakia" & party==96423 & elect.yr==2016))


# This follows the recommendation in Lowe et al. (2011), and is the
#   same procedure used in Barth et al. (2015)
cmp <- cmp %>% 
  mutate(rileemph = log(per104 + 0.5) + log(per201 + 0.5) + log(per203 + 0.5) +
           log(per305 + 0.5) + log(per401 + 0.5) + log(per402 + 0.5) + log(per407 + 0.5) + 
           log(per414 + 0.5) + log(per505 + 0.5) + log(per601 + 0.5) + log(per603 + 0.5) + 
           log(per605 + 0.5) + log(per606 + 0.5) + log(per702 + 0.5) - log(per103 + 0.5) - 
           log(per105 + 0.5) - log(per106 + 0.5) - log(per107 + 0.5) - log(per403 + 0.5) - 
           log(per404 + 0.5) - log(per406 + 0.5) - log(per412 + 0.5) - log(per413 + 0.5) - 
           log(per504 + 0.5) - log(per506 + 0.5) - log(per701 + 0.5) - log(per202 + 0.5) - 
           log(per604 + 0.5)) %>% 
  dplyr::select(countryname, elect.yr, partyname, party, pervote,
                rileemph) %>% 
  rename(country = countryname,
         year = elect.yr)

# Add missing vote shares
# http://electionresources.org/br/president.php?election=2018
cmp$pervote[cmp$country=="Brazil" & cmp$party==180230] <- 29.28
cmp$pervote[cmp$country=="Brazil" & cmp$party==180251] <- 12.47
cmp$pervote[cmp$country=="Brazil" & cmp$party==180310] <- 4.76
cmp$pervote[cmp$country=="Brazil" & cmp$party==180620] <- 46.03
# http://electionresources.org/cl/deputies.php?election=2017
cmp$pervote[cmp$country=="Chile" & cmp$party==155980] <- 1.75
# http://www.dik.co.me/izbori 2016/Konacni rezultati.pdf (only through WebArchive)
cmp$pervote[cmp$country=="Montenegro" & cmp$party==91610] <- 0.40
# Can't identify the second party with a missing vote share in Montenegro
# Have to adjust the vote share for one of the parties in North Macedonia
cmp$pervote[cmp$country=="North Macedonia" & cmp$party==89221] <- 37.87
# The other Macedonian party can't be identified
# The Spanish party with a missing vote share can't be identified

polarDF <- cmp %>% 
  na.omit() %>% 
  group_by(country) %>% 
  summarise(polar_rile = sum(((rileemph - mean(rileemph, na.rm = TRUE))^2)*(pervote/100)))
rm(cmp)

# Correct country names (to be used in merging)
polarDF <- polarDF %>% 
  mutate(country_id = countrycode(country,
                                  origin = "country.name",
                                  destination = "iso3c")) %>% 
  dplyr::select(-country)

write.csv(polarDF,
          file = "./saved/polarization_df.csv",
          row.names = FALSE)
```

* Inter-personal trust from World Values Surveys (*Our World in Data*)

```{r read-owid-data-1, eval=FALSE}
# Data comes from "Our World in Data". Expressed as the % of respondents who
# believe "most people can be trusted", when given the option between this and
# "you can't be too careful"
trust1DF <- read.csv(file = "./raw/OurWorldInData-trust-in-others.csv",
                     header = TRUE)

# Cleaning data
# Also restricting to observations starting with 2009
trust1DF <- trust1DF %>% 
  rename(country = Entity,
         ccode = Code,
         year = Year,
         trust_people = `Trust.in.others....`) %>% 
  dplyr::select(-ccode) %>% 
  group_by(country) %>% 
  filter(year == max(year, na.rm = TRUE)) %>% 
  ungroup() %>% 
  filter(year>=2009) %>% 
  mutate(country_id = countrycode(country,
                                  origin = "country.name",
                                  destination = "iso3c")) %>% 
  filter(!(country %in% c("Burkina Faso", "Mali", "Zambia"))) %>% 
  dplyr::select(-c(country, year))
# This removes 3 countries for which more recent observations can be
# found in Round 5 of the Afrombarometer.


# Read in the most recent data from the Afrobarometer that contains
# a trust item
# afro5DF <- read.spss(file = "../11-Data-repository/71-AB/merged_r5_data_Jul2015.sav",
#                      to.data.frame = TRUE,
#                      use.value.labels = TRUE)
# afro5DF <- afro5DF %>% 
#   dplyr::select(COUNTRY, Q87, withinwt) %>% 
#   mutate(COUNTRY = as.character(COUNTRY),
#          Q87 = as.character(Q87),
#          COUNTRY = if_else(COUNTRY=="Cote d’Ivoire", "Cote d'Ivoire", COUNTRY))
# 
# write.csv(afro5DF,
#           file = "./raw/afrobarometer5_df.csv",
#           row.names = FALSE)

afro5DF <- read.csv(file = "./raw/afrobarometer5_df.csv",
                    header = TRUE)

# This will also remove the countries for which we have a more recent observation 
# in the WVS
afrosumDF <- afro5DF %>% 
  rename(country = COUNTRY,
         trust = Q87,
         weight = withinwt) %>% 
  filter(!(country %in% c("Algeria", "Egypt", "Ghana", "Morocco", "Nigeria",
                          "South Africa", "Tunisia", "Zimbabwe"))) %>% 
  mutate(trust = case_when(trust %in% c("Don't know","Missing") ~ NA_real_,
                           trust=="Must be very careful" ~ 0,
                           trust=="Most people can be trusted" ~ 1),
         trust = as.numeric(trust)) %>%
  na.omit() %>% 
  group_by(country) %>% 
  summarise(trust_people = weighted.mean(trust, weight)*100) %>% 
  ungroup()
rm(afro5DF)

# Produce needed columns to allow for merging with the "Our World in Data"
# data set
afrosumDF <- afrosumDF %>% 
  mutate(country_id = countrycode(country,
                                  origin = "country.name",
                                  destination = "iso3c")) %>% 
  dplyr::select(-country) %>% 
  dplyr::select(trust_people, country_id)

trust1DF <- rbind(trust1DF, afrosumDF)
rm(afrosumDF)

write.csv(trust1DF,
          file = "./saved/trust_people_df.csv",
          row.names = FALSE)
```


* Government trust from World Values Surveys

```{r read-wvs-data, eval=FALSE}

# wvsDF <- readRDS("~/OneDrive/11-Data-repository/19-WVS/WVS_1981_2016_r_v20180912.rds")
load("raw/WVS_1981_2016_r_v20180912_snip.Rdata")

# The item (E069_11) gives respondents a 4-point scale for self-rating their
# confidence in the government: 1=a great deal; 2=quite a lot; 3=not very
# much; 4=none at all.
# The variable we use in the models represents the share of respondents that
# reported trusting the government "a great deal" or "quite a lot"

trust2DF <- trust2DF %>%  
  rename(ccode = S003,
         year = S020,
         trust_gov = E069_11) %>% 
  mutate(trust_gov = as.numeric(trust_gov),
         trust_gov = case_when(trust_gov <= -1 ~ NA_real_,
                               trust_gov %in% c(3,4) ~ 0,
                               trust_gov %in% c(1,2) ~ 1)) %>% 
  group_by(ccode, year) %>% 
  summarise(trust_gov = mean(trust_gov, na.rm = TRUE)*100) %>% 
  ungroup() %>%
  group_by(ccode) %>% 
  filter(year == max(year, na.rm = TRUE)) %>% 
  ungroup() %>% 
  filter(year >= 2009) %>% 
  na.omit() %>% 
  mutate(country_id = countrycode(ccode,
                                  origin = "wvs",
                                  destination = "iso3c")) %>% 
  dplyr::select(-c(ccode, year)) %>% 
  filter(!(country_id %in% c("ARG","BRA","CHL","COL",
                             "ECU","PER","MEX","URY")))

# Add data from the most recent version of the Latin Barometer (2018)
# The data set is larger, so a reduced version of it can be saved in
# the "raw/" subfolder
# latbaroDF <- readRDS(file = "../../11-Data-repository/72-LAPOP/Latinobarometro_2018_Esp_R_v20190303.rds")
# latbaroDF <- latbaroDF %>%
#   dplyr::select(IDENPA, P15STGBSC.E, WT)
# save(latbaroDF, file = "./raw/latinobarometer_2018.Rdata")

load("./raw/latinobarometer_2018.Rdata")
latbaroDF <- latbaroDF %>% 
  rename(ccode = IDENPA,
         trust_gov = P15STGBSC.E,
         weight = WT) %>% 
  mutate(trust_gov = as.numeric(trust_gov),
         trust_gov = case_when(trust_gov %in% c(-1,-2) ~ NA_real_,
                               trust_gov %in% c(3,4) ~ 0,
                               trust_gov %in% c(1,2) ~ 1)) %>% 
  group_by(ccode) %>% 
  summarise(trust_gov = weighted.mean(trust_gov, 
                                      weight, 
                                      na.rm = TRUE)*100) %>% 
  ungroup() %>% 
  na.omit() %>% 
  mutate(country_id = countrycode(ccode,
                                  origin = "wvs",
                                  destination = "iso3c")) %>%
  dplyr::select(-c(ccode))

trust2DF <- rbind(trust2DF, latbaroDF)
rm(latbaroDF)

write.csv(trust2DF,
          "./saved/trust_gov_df.csv",
          row.names = FALSE)
```



## Populism

```{r get-populism-data, eval = FALSE}
# CURRENTLY NOT USED
## DOI NOT CORRECT BELOW
download.file("https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/LFTQEZ/", 
              "./saved/populism.zip",
              mode = "wb",
              quiet = TRUE)
 load("saved/GPD_20190625.RData")

```
 
```{r}
# Populism in power
# https://institute.global/policy/high-tide-populism-power-1990-2020

populism_df <- WDI::WDI_data$country %>% 
  data.frame() %>% 
  select(iso3c, country) %>%
  mutate(electoral_pop = 1*(iso3c %in% c("USA", "MEX", "NIC", "VEN", 
                                         "BRA", "POL", "HUN", "CZE", 
                                         "ITA", "SRB", "BGR", "BLR", 
                                         "TUR", "ISR", "IND", "LKA", 
                                         "PHL"))) %>% 
  dplyr::select(iso3c, electoral_pop) %>% 
  rename(country_id = iso3c) %>% 
  mutate(country_id = if_else(country_id=="XKX", "RKS", country_id))
nrow(populism_df[populism_df$electoral_pop==1,]) == 17

write.csv(populism_df,
          "./saved/populism_df.csv",
          row.names = FALSE)
```

## Index of federalism

```{r read-dpi-data, eval=download$dpi_data}
dpiDF <- readstata13::read.dta13(file = "./raw/DPI2017_Jan2018.dta",
                    convert.factors = FALSE)

dpiDF <- dpiDF %>%
  dplyr::select(countryname, year, auton, muni,
                state, author, stconst, checks,
                polariz, pr) %>%
  rename(checksVeto = checks,
         polarizVeto = polariz) %>% 
  filter(!is.na(year)) %>% 
  group_by(countryname) %>%
  filter(year == max(year, na.rm = TRUE)) %>%
  ungroup()

# Variable standardization for index of federalism
.std_fun <- function(x) {
  (x - mean(x, na.rm = TRUE))/sd(x, na.rm = TRUE)
}

dpiDF <- dpiDF %>% 
  mutate_at(vars(auton:stconst),
            funs(recode(., `0`=0, `1`=1, `2`=2, .default=NA_real_))) %>% 
  mutate(auton.std = .std_fun(auton),
         muni.std = .std_fun(muni),
         state.std = .std_fun(state),
         author.std = .std_fun(author),
         senate.std = .std_fun(stconst)) %>% 
  dplyr::select(-c(auton, muni, state, author, stconst))

# The last 2 items have 114 and 115 missing values out of 181 countries.
# Select only countries which have 3 valid observations for the measurements
dpiDF$missno <- apply(dpiDF[ ,6:10], 1, function(x) sum(is.na(x)))

fedDF <- dpiDF %>% 
  filter(missno <= 2) %>% 
  mutate(federal_ind = rowMeans(dplyr::select(.,ends_with(".std")), na.rm = TRUE)) %>% 
  filter(year > 2015) %>% 
  dplyr::select(-c(year, pr:missno))

# Generate country code and clean country names
fedDF <- fedDF %>% 
  filter(!(countryname %in% c("Turk Cyprus","GDR","Yemen (PDR)",
                              "Yugoslavia","Soviet Union","Yemen (AR)"))) %>%
  rename(country = countryname) %>% 
  mutate(country = recode(country,
                          "Cent. Af. Rep." = "Central African Republic",
                          "PRC" = "China",
                          "Dom. Rep." = "Dominican Republic",
                          "ROK" = "Republic of Korea",
                          "PRK" = "North Korea",
                          "S. Africa" = "South Africa",
                          "P. N. Guinea" = "Papua New Guinea")) %>% 
  mutate(GeoId = countrycode(country,
                             origin = "country.name",
                             destination = "iso2c")) %>% 
  dplyr::select(-c(country, checksVeto, polarizVeto))  %>% 
  mutate(country_id = countrycode(GeoId,
                                    origin = "iso2c",
                                    destination = "iso3c")) %>% 
  dplyr::select(-GeoId)

write.csv(fedDF,
          file = "./saved/federalism_df.csv",
          row.names = FALSE)
rm(fedDF)
```

```{r read-dpi-data2, eval=download$dpi_data}
# Follow the same procedure for checks and balances indicators
checksDF <- dpiDF %>% 
  dplyr::select(countryname, checksVeto, polarizVeto, year) %>% 
  filter(year > 2015) %>% 
  dplyr::select(-year)

# Generate country code and clean country names
checksDF <- checksDF %>% 
  filter(!(countryname %in% c("Turk Cyprus","GDR","Yemen (PDR)",
                              "Yugoslavia","Soviet Union","Yemen (AR)"))) %>% 
  rename(country = countryname) %>% 
  mutate(country = recode(country,
                          "Cent. Af. Rep." = "Central African Republic",
                          "PRC" = "China",
                          "Dom. Rep." = "Dominican Republic",
                          "ROK" = "Republic of Korea",
                          "PRK" = "North Korea",
                          "S. Africa" = "South Africa",
                          "P. N. Guinea" = "Papua New Guinea")) %>% 
  mutate(country_id = countrycode(country,
                                  origin = "country.name",
                                  destination = "iso3c")) %>% 
  dplyr::select(-c(country)) %>% 
  rename(checks_veto = checksVeto,
         polariz_veto = polarizVeto) %>% 
  mutate(checks_veto = ifelse(checks_veto == -999, NA, checks_veto),
         polariz_veto = ifelse(polariz_veto == -999, NA, checks_veto))

write.csv(checksDF, file = "./saved/dpi_checks_df.csv", row.names = FALSE)
#rm(checksDF)
```

```{r read-dpi-data3, eval=download$dpi_data}
prDF <- dpiDF %>% 
  dplyr::select(countryname, pr)

# Generate country code and clean country names
prDF <- prDF %>% 
  filter(!(countryname %in% c("Turk Cyprus","GDR","Yemen (PDR)",
                              "Yugoslavia","Soviet Union","Yemen (AR)"))) %>% 
  rename(country = countryname) %>% 
  mutate(country = recode(country,
                          "Cent. Af. Rep." = "Central African Republic",
                          "PRC" = "China",
                          "Dom. Rep." = "Dominican Republic",
                          "ROK" = "Republic of Korea",
                          "PRK" = "North Korea",
                          "S. Africa" = "South Africa",
                          "P. N. Guinea" = "Papua New Guinea")) %>%
  mutate(pr = as.numeric(pr),
         pr = if_else(pr==-999, NA_real_, pr)) %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  dplyr::select(-c(country)) 

write.csv(prDF,
          file = "./saved/pr_df.csv",
          row.names = FALSE)
#rm(prDF)
```


# Electoral pressure

## Electoral pressure data from WIKI

```{r get-datawiki, eval=download$wiki_data}
#Loading Wikipedia URL containing worldwide upcoming election data
url <- "https://en.wikipedia.org/wiki/List_of_next_general_elections"

# Scrape continental election tables from Wikipedia and create single 
# world election data frame
##Define scraping function
.scrape_table = function(tableindex){
  url <- "https://en.wikipedia.org/wiki/List_of_next_general_elections"
  tbls_ls <- url %>%
    read_html() %>%
    html_nodes("table") %>%
    .[tableindex] %>%
    html_table(fill = TRUE)
  continent <- tbls_ls[[1]]
  rm(tbls_ls)
  colnames(continent) <- c("country","parlTerm","parlLast","parlNext",
                        "presTerm","presLast","presNext","fairness",
                        "inPower")
  continent <- continent %>% 
    slice(-1) %>% 
    dplyr::select(country, parlLast, parlNext,
                presLast, presNext)
 return(continent)
}

## Call scraping function for Africa, Americas, Asia and Oceania
africaDF  = .scrape_table(2) 
americaDF = .scrape_table(3) 
asiaDF    = .scrape_table(4) 
oceaniaDF = .scrape_table(6)

## Europe
tbls_ls <- url %>%
  read_html() %>%
  html_nodes("table") %>%
  .[5] %>%
  html_table(fill = TRUE)
europeDF <- tbls_ls[[1]]
rm(tbls_ls)
colnames(europeDF) <- c("country","parlTerm","parlLast","parlNext",
                        "presTerm","presLast","presNext","fairness",
                        "inPower", "status")
europeDF <- europeDF %>% 
  slice(-1) %>% 
  dplyr::select(country, parlLast, parlNext,
                presLast, presNext)

# Create single world election dataframe 
worldDF <- rbind(africaDF, americaDF, asiaDF, europeDF, oceaniaDF)
# Drop Continental dataframes
rm(africaDF, americaDF, asiaDF, europeDF, oceaniaDF)

# Create countrycodes and deal with missing values
worldDF <- worldDF %>% 
  filter(!(country=="European Union")) %>% 
  mutate(country_id = countrycode(country,
                                  origin = "country.name",
                                  destination = "iso3c")) %>% 
  mutate(country_id = if_else(country=="Kosovo", "RKS", country_id)) %>%
  mutate_at(vars(parlLast:presNext), as.character) %>% 
  mutate(parlLast = if_else(parlLast %in% c("","N/A"), NA_character_, parlLast),
         parlNext = if_else(parlNext %in% c("","N/A"), NA_character_, parlNext),
         presLast = if_else(presLast %in% c("","N/A"), NA_character_, presLast),
         presNext = if_else(presNext %in% c("","N/A"), NA_character_, presNext)) %>% 
  mutate(presNext = if_else(presNext=="2020[6][7]", "2020", presNext))

# It's a pretty standardized format, so it can be done quite quickly,
# though the next dates would require considerably more manual fiddling.
# tempDF <- worldDF %>% 
#   mutate(YparlLast = str_sub(parlLast, start = -4),
#          MparlLast = str_match(parlLast, " (.*?) ")[ ,2],
#          DparlLast = str_sub(parlLast, start = 1, end = 2))

# Define function to split day, month and year for next and last 
# parliament elections to clean dates
.split_dates = function(datecolumn){
resultSplitted = as.data.frame(matrix(-1, nrow= length(datecolumn), ncol = 4))
splitted = sapply(datecolumn, str_split, pattern = " ")

for(x in 1:length(datecolumn)){
  if(length(splitted[[x]]) == 1){
    resultSplitted[x, 3] = splitted[[x]][1]
  }
  if(length(splitted[[x]]) == 2){
    resultSplitted[x, 2]= splitted[[x]][1]
    resultSplitted[x, 3] = splitted[[x]][2]
  }
  if(length(splitted[[x]]) == 3){
    resultSplitted[x, 1] = splitted[[x]][1]
    resultSplitted[x, 2]= splitted[[x]][2]
    resultSplitted[x, 3] = splitted[[x]][3]
  }
  resultSplitted[x, 4] = worldDF$country_id[x]
}
  return(resultSplitted)
}

# Call function to split day, month and year for next and last parliament 
# elections to clean dates

# PARLIMENTARY
# Create dataframe for splitted next PARLIMENTARY election date
resultsplittedNext = .split_dates(worldDF$parlNext)
colnames(resultsplittedNext) <- c("DayNext","MonthNext",
                                  "YearNext","country_id")

# Create dataframe for splitted last PARLIMENTARY election date
resultsplittedLast = .split_dates(worldDF$parlLast)
colnames(resultsplittedLast) <- c("DayLast","MonthLast",
                                  "YearLast","country_id")

# Merge dataframes for splitted next and last PARLIMENTARY election 
# date to prepare cleaning
wikiParlDF <- left_join(resultsplittedNext, 
                        resultsplittedLast, 
                        by = "country_id")
rm(resultsplittedLast,resultsplittedNext)

# PRESIDENTIAL
# Create dataframe for splitted next PRESIDENTIAL election date
resultsplittedNext = .split_dates(worldDF$presNext)
colnames(resultsplittedNext) <- c("DayNext","MonthNext",
                                  "YearNext","country_id")

# Create dataframe for splitted last PRESIDENTIAL election date
resultsplittedLast = .split_dates(worldDF$presLast)
colnames(resultsplittedLast) <- c("DayLast","MonthLast",
                                  "YearLast","country_id")

# Merge dataframes for splitted next and last PRESIDENTIAL election 
# date to prepare cleaning
wikiPresDF <- left_join(resultsplittedNext, 
                        resultsplittedLast, 
                        by = "country_id")
rm(resultsplittedLast, resultsplittedNext, worldDF)


# Deal with dates: 
# (1) take day and month of last election date if both of these
# are missing
# (2) take day of last election if only day is missing
# (3) set day to the 15th if there is a mismatch between last
# election month and future election month (e.g. parliamentary
# elections in CAF)

# Solve missing month and day
# Correct manually a few cases (like Libya) where election timeline
# is not clear at all.
wikiParlDF <- wikiParlDF %>% 
  mutate(MonthNext = if_else(MonthNext==-1, MonthLast, MonthNext),
         DayNext = if_else(DayNext==-1, DayLast, DayNext),
         DayNext = if_else(MonthNext != MonthLast, "15", DayNext)) %>% 
  mutate(DayNext = if_else(is.na(YearNext), NA_character_, DayNext),
         MonthNext = if_else(is.na(YearNext), NA_character_, MonthNext))

# Convert to dates
wikiParlDF <- wikiParlDF %>% 
  unite("nextWIKIparl", DayNext:YearNext, sep = "-") %>% 
  mutate(nextWIKIparl = if_else(nextWIKIparl=="NA-NA-NA", NA_character_, nextWIKIparl),
         nextWIKIparl = as.Date(nextWIKIparl, "%d-%B-%Y"))

# Follow the same procedure for presidential elections
wikiPresDF <- wikiPresDF %>% 
  mutate(MonthNext = if_else(MonthNext==-1, MonthLast, MonthNext),
         DayNext = if_else(DayNext==-1, DayLast, DayNext),
         DayNext = if_else(MonthNext != MonthLast, "15", DayNext)) %>% 
  mutate(DayNext = if_else(is.na(YearNext), NA_character_, DayNext),
         MonthNext = if_else(is.na(YearNext), NA_character_, MonthNext))

# Convert to dates
wikiPresDF <- wikiPresDF %>% 
  unite("nextWIKIpres", DayNext:YearNext, sep = "-") %>% 
  mutate(nextWIKIpres = if_else(nextWIKIpres=="NA-NA-NA", NA_character_, nextWIKIpres),
         nextWIKIpres = as.Date(nextWIKIpres, "%d-%B-%Y"))

# Merge parliament and president election dataframes into one wiki dataframe
NextWikiElectionDF <- left_join(select(wikiPresDF, nextWIKIpres:country_id), 
                                select(wikiParlDF, nextWIKIparl:country_id), 
                                by = "country_id")
rm(wikiParlDF, wikiPresDF)
```

## Electoral pressure data from IEFS 

```{r get-dataiefs, eval=download$iefs_data}
# Read in IEFS data and costruct contrycodes
IEFS <- read.csv(file = "./raw/IEFS_election_dates_2020.csv")

IEFS <- IEFS %>% 
  mutate(country_id = countrycode(Country,
                                  origin = "country.name",
                                  destination = "iso3c"))





# Create separate columns for presidential, parliament and senate elections 
# for confirmed election dates only
# Restrict analysis to confirmed election dates for data quality reasons and 
# create separate df for presidential, parliamentary and senate elections
# Drop referenda
# Drop Mali's 2nd parliamentary round, under the assumption that the period
# between the two rounds is too short for governments to be able to hope
# for a change in popularity.
IEFS <- IEFS %>% 
  filter(Status=="Confirmed") %>% 
  filter(Body != "Referendum") %>% 
  dplyr::select(-c(Status, Month)) %>% 
  mutate(Body = as.character(Body),
         Body = if_else(Body=="President Round 2", "President", Body),
         Body = if_else(Body=="Governor ", "President", Body)) %>%
  filter(Body != "Parliament Round 2") %>% 
  filter(!(country_id=="KIR" & Date=="4/15/2020")) %>% 
  filter(Body != "NVD") %>% 
  pivot_wider(id_cols = c(Country, country_id),
              names_from = Body,
              values_from = Date)

# Bring dates into dates format
IEFS <- IEFS %>% 
  mutate(Parliament = as.Date(Parliament, "%m/%d/%Y"),
         President = as.Date(President, "%m/%d/%Y"),
         Senate = as.Date(Senate, "%m/%d/%Y"))

# The following filter is applied because the electoral pressure measure 
# is constructed as the distance of the next election from the WHO 
# "pandemic" declaration on March 11, 2020. Negative distance do not make 
# sense and are excluded by only taking in days after March 11, 2020
parlDF <- IEFS %>% 
  filter(Parliament > "2020-03-11") %>% 
  dplyr::select(country_id, Parliament)
presDF <- IEFS %>% 
  filter(President > "2020-03-11") %>% 
  dplyr::select(country_id, President)
senateDF <- IEFS %>% 
  filter(Senate > "2020-03-11") %>% 
  dplyr::select(country_id, Senate)

# A tricky merge, since some elections are captured in the presidential
# data set, but not in the parliamentary one, and vice-versa
mergedNextElectionDF <- full_join(parlDF,
                                  presDF,
                                  by = "country_id")
# For second one it's fine to do "left_join()"
mergedNextElectionDF <- left_join(mergedNextElectionDF,
                                  senateDF,
                                  by = "country_id")
# Bring in Wikipedia data as well
mergedNextElectionDF <- full_join(mergedNextElectionDF,
                                  NextWikiElectionDF,
                                  by = "country_id")
rm(IEFS, parlDF, presDF, senateDF, NextWikiElectionDF)
```

## Electoral pressure data from IPU

```{r get-dataipu, eval=download$ipu_data}
# Read in IEFS data and construct country codes
IPU <-  read.csv(file = "./raw/elections-export.csv", skip = 10) %>% 
  mutate(country_id = countrycode(sourcevar = Country,
                                  origin = "country.name",
                                  destination = "iso3c"))

# Create data frame for upper chambers with election (include: direct 
# and indirect elections)
IPUsena <- IPU %>% 
  dplyr::select(-c(ISO.Code, Structure.of.parliament, Country)) %>% 
  rename(type = 2,
         designation = 3,
         pastIPUsena = 4,
         nextIPUsena = 5) %>% 
  filter(type=="Upper chamber") %>% 
  filter(designation %in% c("Directly elected","Indirectly elected"))

IPUparl <- IPU %>% 
  dplyr::select(-c(ISO.Code, Structure.of.parliament, Country)) %>% 
  rename(type = 2,
         designation = 3,
         pastIPUparl = 4,
         nextIPUparl = 5) %>% 
  filter(type=="Lower chamber") %>% 
  filter(designation %in% c("Directly elected","Indirectly elected"))

# Producing a date column out of these
IPUparl <- IPUparl %>% 
  mutate(nextIPUparl = if_else(nextIPUparl=="-", NA_character_, nextIPUparl)) %>% 
  mutate(nextIPUparl = as.Date(nextIPUparl, "%d %b %Y")) %>% 
  dplyr::select(country_id, nextIPUparl)

IPUsena <- IPUsena %>% 
  mutate(nextIPUsena = if_else(nextIPUsena=="-", NA_character_, nextIPUsena)) %>% 
  mutate(nextIPUsena = as.Date(nextIPUsena, "%d %b %Y")) %>% 
  dplyr::select(country_id, nextIPUsena)



## Full merge selected variables into main data frame to maintain all countries 
## of both data sets
mergedNextElectionDF <- full_join(mergedNextElectionDF, 
                                  IPUparl, 
                                  by = "country_id")
mergedNextElectionDF <- full_join(mergedNextElectionDF, 
                                  IPUsena, 
                                  by = "country_id")
rm(IPUsena, IPUparl)


## Deal with other selection procedures for members of the upper chambers 
## than election: in particular appointments
IPUnosena <- IPU %>% 
  dplyr::select(-c(ISO.Code, Structure.of.parliament, Country)) %>% 
  rename(type = 2,
         designation = 3,
         pastAppointIPUsena = 4,
         nextAppointIPUsena = 5) %>% 
  filter(type=="Upper chamber") %>% 
  filter(designation %in% c("Appointed","Other")) %>% 
  dplyr::select(country_id) %>% 
  mutate(noelectionIPUsena = 1)

IPUnoparl <- IPU %>% 
  dplyr::select(-c(ISO.Code, Structure.of.parliament, Country)) %>% 
  rename(type = 2,
         designation = 3,
         pastAppointIPUparl = 4,
         nextAppointIPUparl = 5) %>% 
  filter(!(type=="Upper chamber")) %>% 
  filter(designation %in% c("Appointed","Other")) %>% 
  dplyr::select(country_id) %>% 
  mutate(noelectionIPUparl = 1)



mergedNextElectionDF <- full_join(mergedNextElectionDF, 
                                  IPUnoparl, 
                                  by = "country_id")
mergedNextElectionDF <- full_join(mergedNextElectionDF, 
                                  IPUnosena, 
                                  by = "country_id")
rm(IPUnoparl,IPUnosena, IPU)
```

# Create electoral pressure measures from IPU, Wiki and IEFS data sets

```{r, eval=FALSE}
# Create single next election date for parliament, senate and president using 
# IPU, WIKI and IEFS data taking into account data quality (IEFS > IPU > WIKI) 
mergedNextElectionDF <- mergedNextElectionDF %>% 
  rename(nextparl = Parliament,
         nextpres = President,
         nextsena = Senate) %>% 
  mutate(nextparl = if_else(is.na(nextparl), nextIPUparl, nextparl),
         nextparl = if_else(is.na(nextparl), nextWIKIparl, nextparl),
         nextpres = if_else(is.na(nextpres), nextWIKIpres, nextpres),
         nextsena = if_else(is.na(nextsena), nextIPUsena, nextsena)) %>% 
  dplyr::select(-c(nextIPUparl, nextWIKIparl, nextWIKIpres, nextIPUsena))

# Create raw distance measures for next election date for parliament, senate 
# and president 
# This is constructed by using the distance of the next election from the 
# WHO "pandemic" declaration on March 11, 2020
# We exclude negative distance for past elections (elections prior to 
# March 11,2020).
# We set days to next election to 2000 for countries with no election 
# (for distance measure)
mergedNextElectionDF <- mergedNextElectionDF %>% 
  mutate(dist_parlm = as.numeric(nextparl - as.Date("2020-03-11")),
         dist_presid = as.numeric(nextpres - as.Date("2020-03-11")),
         dist_senate = as.numeric(nextsena - as.Date("2020-03-11"))) %>% 
  mutate(dist_parlm = if_else(dist_parlm < 0, NA_real_, dist_parlm),
         dist_presid = if_else(dist_presid < 0, NA_real_, dist_presid),
         dist_senate = if_else(dist_senate < 0, NA_real_, dist_senate)) %>% 
  mutate(distancenoelection = if_else(is.na(dist_parlm) &
                                        is.na(dist_presid) &
                                        is.na(dist_senate), 2000, NA_real_))

### Measure Creation ###

# 1. Create single distance measure for next election without 
# discriminating between president, senate, parliament elections
mergedNextElectionDF$dist_anyelec <- apply(mergedNextElectionDF[,c(7:10)], 
                                           1, 
                                           min, na.rm = TRUE)

# 2. Create categorical electoral cycle measure (factor) with 
# three categories
mergedNextElectionDF$electoralcycle <- cut(mergedNextElectionDF$dist_anyelec, 
                                           c(0, 365, 730, 2000))
# Categories: 1 = next election within one year; 
# 2 = next election within two years; 
# 3 = next election not within two years time from March 11, 2020 

# 3. Create electoral pressure measure with range [0,1] with 1 being 
# high electoral pressure and zero being no electoral pressure
mergedNextElectionDF$elect_pressure <- 1/ (apply(mergedNextElectionDF[,c(7:9)], 
                                                 1, 
                                                 min, na.rm = TRUE))
# Note: Here we deal with the creation of infinity values due to the 
# application of the min function with dividing 1 by the days to 
# next election
# -> If a state has no upcoming election scheduled, infinity is created 
# and through our treatment, electoral pressure is set to be 0  

#Select columns of data frame relevant for the analysis
mergedNextElectionDF <- mergedNextElectionDF %>% 
  dplyr::select(dist_senate, dist_presid, dist_parlm, 
                dist_anyelec, elect_pressure, country_id)

write.csv(mergedNextElectionDF, 
          file = "./saved/electoral_pressure_df.csv", 
          row.names = FALSE)
```


# Adding Community Mobility Report data

```{r get-datamob, eval=download$mobility_data}

mobilityDF <- read.csv(url("https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv")) %>% 
  filter(sub_region_1 == "") %>% 
  rename(retail = retail_and_recreation_percent_change_from_baseline, 
         grocery = grocery_and_pharmacy_percent_change_from_baseline, 
         parks = parks_percent_change_from_baseline,
         transit = transit_stations_percent_change_from_baseline, 
         work = workplaces_percent_change_from_baseline,
         residential = residential_percent_change_from_baseline) %>% 
  mutate(country_id = countrycode(country_region_code,
                                    origin = "iso2c",
                                    destination = "iso3c"), 
         mobility_index = (retail + grocery + parks + transit + work) / 5 ) %>%  # Creating mobility index
  mutate(country_id = if_else(is.na(country_id), "NAM", country_id)) %>% 
  dplyr::select(- sub_region_1, - sub_region_2, - country_region_code, - country_region) %>%
  mutate(date = as.Date(date))

## Fill missing values down

mobilityDF <- mobilityDF %>% 
  arrange(country_id, date) %>% 
  group_by(country_id) %>% 
  fill(mobility_index, .direction = 'down')

## save

write.csv(mobilityDF,  file = "./saved/mobility_level_df.csv", row.names = FALSE)
```

# Government Left-Right placement

In the *ParlGov* source, every party on the spectrum is assigned a Left-Right ideological score. Parties that are in the governing coalition are also identified. With these 2 pieces of information we can construct an ideological placement for the cabinet. Though a few choices can be made, a good start is to construct the government placement as the weighted average of the positions of the parties in government, using their seat shares as weights.^[Other alternatives exist. If we assume that a party withdrawing from the government coalition can bring the entire government down, then an unweighted average of party positions might be a better indication of the government's policy orientation. We might also consider that the party which holds the portfolio of prime minister has an disproportionate influence on the orientation of the government, leading to a higher weight for this party. A similar logic could be applied for the party which holds the Economy or Finance portfolio.]

```{r get-data, eval=download$parlgov_data}
parlgovDF <- read.csv(url("http://www.parlgov.org/static/data/development-cp1252/view_cabinet.csv"),
                      header = TRUE)

parlgovDF <- parlgovDF %>% 
  dplyr::select(country_name, election_date, start_date, cabinet_name, 
                cabinet_party, seats, party_name_english, left_right) %>% 
  rename(country = country_name,
         election_date = election_date,
         start.date = start_date,
         cab.name = cabinet_name,
         cabinet = cabinet_party,
         p_name = party_name_english,
         lr_pos = left_right) %>% 
  mutate(year = as.numeric(str_sub(election_date, start = 1, end = 4))) %>% 
  filter(year >= 2008) %>%

# Keep only the most recent election for each country; I confirmed that
# it's truly the most recent election from Adam Carr's Electoral Archive.
  mutate(start.date = as.Date(start.date)) %>% 
  group_by(country) %>% 
  filter(start.date == max(start.date, na.rm = TRUE)) %>% 
  ungroup() %>% 
  dplyr::select(-c(start.date, cab.name, year)) %>% 
  filter(cabinet==1) %>% 
  dplyr::select(-cabinet) %>% 
  filter(country != "Italy") %>% 
  na.omit()


# Save this version of the data - we'll need it to construct 2 different
# indicators in the data set.
write.csv(parlgovDF,
          file = "./saved/parlgov_cabinets_df.csv",
          row.names = FALSE)
```


```{r cabinet-position}

parlgovDF <- read.csv(file = "./saved/parlgov_cabinets_df.csv")

# Compute the position of the government
positionDF <- parlgovDF %>% 
  group_by(country) %>% 
  summarise(pos_gov_lr = sum(seats*lr_pos)/sum(seats)) %>% 
  mutate(country_id = countrycode(country,
                                    origin = "country.name",
                                    destination = "iso3c")) %>% 
  dplyr::select(-country)


write.csv(positionDF,
          file = "./saved/gov_position_df.csv",
          row.names = FALSE)

```

Higher scores on this variable denote a more rightward position of the government.


## Women leaders

Taken from https://en.wikipedia.org/wiki/List_of_elected_and_appointed_female_heads_of_state_and_government#Elected_or_appointed_female_chief_executives and gives 1 to a state with a woman head of government on 1 Jan 2020. This includes Austria where Brigitte Bierlein left office on 7 Jan and it excludes and Marshall islands where Hilda Heine left on 14 Jan.

```{r get-women-leaders-data, eval = FALSE}
woman_leader_df <- WDI::WDI_data$country %>% 
  data.frame() %>% 
  select(iso3c, country) %>%
  mutate(woman_leader = 1*(iso3c %in% c("AUT", "BEL", "BGD", 
                                        "BOL", "BRB", "DEU", 
                                        "DNK", "FIN", "ISL", 
                                        "MHL", "MMR", "NAM", 
                                        "NOR", "NZL", "SRB", 
                                        "TWN"))) %>% 
  dplyr::select(iso3c, woman_leader) %>% 
  mutate(iso3c = if_else(iso3c=="XKX", "RKS", iso3c)) %>% 
  rename(country_id = iso3c)

nrow(woman_leader_df[woman_leader_df$woman_leader==1,]) == 16

write.csv(woman_leader_df,
          "./saved/woman_leader_df.csv",
          row.names = FALSE)
```

# Strigency of policies index 
```{r, read-stringency-election, eval=download$stringency_data}
#Read in Oxford stringency measures
stringencydf <- read.csv(url("https://raw.githubusercontent.com/OxCGRT/covid-policy-tracker/master/data/OxCGRT_latest.csv")) %>% 
  select(CountryCode , Date, StringencyIndex, C1_School.closing, C2_Workplace.closing, C3_Cancel.public.events, C4_Restrictions.on.gatherings, C5_Close.public.transport, C6_Stay.at.home.requirements, C7_Restrictions.on.internal.movement, C8_International.travel.controls, H1_Public.information.campaigns)  %>% 
  mutate(Date = as.Date(as.character(Date), "%Y%m%d")) %>% 
  rename(country_id = CountryCode, date = Date)

## Fill missing values down

stringencydf <- stringencydf %>% 
  arrange(country_id, date) %>% 
  group_by(country_id) %>% 
  fill(StringencyIndex, .direction = 'down') %>% 
  rename(stringency = StringencyIndex)


write.csv(stringencydf,"./saved/stringency_df.csv",row.names = FALSE)
```
